fix(mobile): keep workspaces whose project the host stopped reporting
Grouping keyed on the raw projectId, but sections are built from the reported
project list — so a workspace pointing at a project that isn't in that list
rendered nowhere at all. Projects and workspaces are separate queries, so this
also swallowed the whole list for the beat where workspaces had arrived and
projects hadn't. Unknown ids now fall into the same bucket as none.

Also: the collapse store learns when it has hydrated, so persisted sections
stop flashing open and snapping shut on a cold start; and the scope chip moves
into its own file per the one-component-per-file rule. All three found by
CodeRabbit on #6512.
