+++ apps/mobile/screens/(authenticated)/(home)/home/components/ScopeBar/ScopeBar.tsx
@@ -1,8 +1,6 @@
-import { ChevronDown } from "lucide-react-native";
-import { Pressable, View } from "react-native";
-import { Text } from "@/components/ui/text";
-import { useTheme } from "@/hooks/useTheme";
+import { View } from "react-native";
 import { HostStatusDot } from "@/screens/(authenticated)/components/HostStatusDot";
+import { Chip } from "./components/Chip";
 
 interface ScopeBarProps {
 	hostName: string | null;
@@ -10,31 +8,6 @@ interface ScopeBarProps {
 	sortLabel: string;
 	onPressHost: () => void;
 	onPressSort: () => void;
-}
-
-function Chip({
-	label,
-	leading,
-	onPress,
-}: {
-	label: string;
-	leading?: React.ReactNode;
-	onPress: () => void;
-}) {
-	const theme = useTheme();
-	return (
-		<Pressable
-			onPress={onPress}
-			accessibilityLabel={label}
-			className="bg-secondary/60 flex-row items-center gap-1.5 rounded-full py-1.5 pl-3 pr-2.5 active:opacity-60"
-		>
-			{leading}
-			<Text className="text-foreground max-w-40 text-xs" numberOfLines={1}>
-				{label}
-			</Text>
-			<ChevronDown size={12} color={theme.mutedForeground} strokeWidth={2.5} />
-		</Pressable>
-	);
 }
 
 /**
