+++ apps/mobile/screens/(authenticated)/(home)/home/components/ScopeBar/components/Chip/index.ts
@@ -1,0 +1,1 @@
+export { Chip } from "./Chip";
